+++ itemset/khhcentroidcounter.go
@@ -31,22 +31,17 @@ func NewKHHCentroidCounter(k int) *KHHCentroidCounter {
     confidence := 0.99;
     seed := int(time.Now().UnixNano() / int64(time.Millisecond));
     var countlist map[int32]int32;
-
-     /* TODO: Need Comparator in Heap. */
-    // cmp := new Comparator<Centroid>() {
-    //     @Override
-    //     public int compare(Centroid n1, Centroid n2) {
-    //         long cn1 = countlist.get(n1.id);// count(n1.id);
-    //         long cn2 = countlist.get(n2.id);// count(n2.id);
-    //         if (cn1 > cn2)
-    //             return +1;
-    //         else if (cn1 < cn2)
-    //             return -1;
-    //         return 0;
-    //     }
-    // };
-
-    priorityQueue := utils.NewPQueue(/*cmp*/);
+    compare := func(n1, n2 types.Centroid) int {
+        cn1 := countlist[n1.id];
+        cn2 := countlist[n2.id];
+        if cn1 > cn2 {
+            return 1;
+        } else if cn1 < cn2 {
+            return -1;
+        }
+        return 0;
+    };
+    priorityQueue := utils.NewPQueue(compare, types.Centroid);
     var frequentItems map[int32]types.Centroid;
     width := int(math.Ceil(2 / epsOfTotalCount));
     depth := int(math.Ceil(-math.Log(float64(1 - confidence)) / math.Log(2)));
@@ -73,44 +68,44 @@ func NewKHHCentroidCounter(k int) *KHHCentroidCounter {
 };
 
 func (this *KHHCentroidCounter) Add(c types.Centroid) {
-	this.count++;
-	count := this.AddLong(c.id, 1);
-    delete(frequentItems, c.id);
-    probed = frequentItems[c.id];
+    this.count++;
+    count := this.AddLong(c.id, 1);
+    delete(this.frequentItems, c.id);
+    probed := this.frequentItems[c.id];
 
-	for _, h := range c.ids {
-		if probed != nil {
-			break;
-		}
-        delete(frequentItems, h);
-		probed = frequentItems[h];
-	}
+    for _, h := range c.ids {
+        if probed != nil {
+            break;
+        }
+        delete(this.frequentItems, h);
+        probed = this.frequentItems[h];
+    }
 
-	if probed == nil {
-		countlist[c.id] = count;
-		frequentItems[c.id] = c;
-		priorityQueue.Add(c);
-	} else {
-		priorityQueue.Remove(probed);
-		probed.UpdateVector(c.Centroid());
-		probed.ids.AddAll(c.ids);
-		frequentItems[probed.id] = probed;
-		countlist[probed.id] = count + 1;
-		priorityQueue.Add(probed);
-	}
+    if probed == nil {
+        this.countlist[c.id] = count;
+        this.frequentItems[c.id] = c;
+        this.priorityQueue.Add(c);
+    } else {
+        this.priorityQueue.Remove(probed);
+        probed.UpdateVector(c.Centroid());
+        probed.ids.AddAll(c.ids);
+        this.frequentItems[probed.id] = probed;
+        this.countlist[probed.id] = count + 1;
+        this.priorityQueue.Add(probed);
+    }
 
-	if priorityQueue.Size() > this.k {
-		removed := priorityQueue.Poll();
-		delete(frequentItems, removed.id);
-		delete(countlist, removed.id);
-	}
+    if this.priorityQueue.Size() > this.k {
+        removed := this.priorityQueue.Poll();
+        delete(this.frequentItems, removed.id);
+        delete(this.countlist, removed.id);
+    }
 };
 
 func (this *KHHCentroidCounter) Hash(item int32, int i) int {
-	var hash uint32 = this.hashA[i] * item;
-	hash += hash >> 32;
-	hash &= PRIME_MODULUS;
-	return int(hash % this.width);
+    var hash uint32 = this.hashA[i] * item;
+    hash += hash >> 32;
+    hash &= PRIME_MODULUS;
+    return int(hash % this.width);
 };
 
 /**
@@ -121,45 +116,45 @@ func (this *KHHCentroidCounter) Hash(item int32, int i) int {
  * @return size of min count bucket
  */
 func (this *KHHCentroidCounter) AddLong(item, count int32) int32 {
-	this.table[0][this.Hash(item, 0)] += count;
+    this.table[0][this.Hash(item, 0)] += count;
     min := int(this.table[0][this.Hash(item, 0)]);
-	for i := 1; i < depth; i++ {
-		this.table[i][this.Hash(item, i)] += count;
-		if this.table[i][this.Hash(item, i)] < min {
+    for i := 1; i < depth; i++ {
+        this.table[i][this.Hash(item, i)] += count;
+        if this.table[i][this.Hash(item, i)] < min {
             min = int(this.table[i][this.Hash(item, i)]);
         }
-	}
-	return min;
+    }
+    return min;
 };
 
 func (this *KHHCentroidCounter) Count(item int32) int32 {
     min := int(this.table[0][this.Hash(item, 0)]);
-	for i := 1; i < this.depth; i++ {
-		if this.table[i][this.Hash(item, i)] < min {
+    for i := 1; i < this.depth; i++ {
+        if this.table[i][this.Hash(item, i)] < min {
             min = int(this.table[i][this.Hash(item, i)]);
         }
-	}
-	return min;
+    }
+    return min;
 };
 
 func (this *KHHCentroidCounter) GetTop() []types.Centroid {
-	if this.topcent != nil {
+    if this.topcent != nil {
         return this.topcent;
     }
     this.topcent = []types.Centroid{};
     this.counts = []int32{};
-	for !this.priorityQueue.IsEmpty() {
+    for !this.priorityQueue.IsEmpty() {
         tmp := this.priorityQueue.Poll();
-		append(this.topcent, tmp);
-		append(this.counts, this.Count(tmp.id));
-	}
-	return this.topcent;
+        append(this.topcent, tmp);
+        append(this.counts, this.Count(tmp.id));
+    }
+    return this.topcent;
 };
 
-func (this *KHHCentroidCounter) GetCounts() [][]float64 {
-	if this.counts != nil {
+func (this *KHHCentroidCounter) GetCounts() []int32; {
+    if this.counts != nil {
         return this.counts;
     }
-	this.GetTop();
-	return this.counts;
+    this.GetTop();
+    return this.counts;
 };